Add write buffer file adaptor.
It's used along with read buffer file adaptor in cases where user's snapshot
file is read/written with some limits, for example it's needed to read or
write at least an aligned bytes, or Integer number of user-defined
objects(such as K-V pairs). File hole is supported by the way.